+++ demo/admin/src/pages/blocks/MediaBlock.tsx
@@ -1,6 +1,5 @@
 import { BlockCategory, createOneOfBlock } from "@comet/blocks-admin";
 import { DamImageBlock, DamVideoBlock, VimeoVideoBlock, YouTubeVideoBlock } from "@comet/cms-admin";
-import * as React from "react";
 import { FormattedMessage } from "react-intl";
 
 export const MediaBlock = createOneOfBlock({