+++ demo/site/src/util/StyledComponentsRegistry.tsx
@@ -1,10 +1,10 @@
 "use client";
 import theme from "@src/theme";
 import { useServerInsertedHTML } from "next/navigation";
-import React, { useState } from "react";
+import { PropsWithChildren, useState } from "react";
 import { ServerStyleSheet, StyleSheetManager, ThemeProvider } from "styled-components";
 
-export default function StyledComponentsRegistry({ children }: { children: React.ReactNode }) {
+export default function StyledComponentsRegistry({ children }: PropsWithChildren) {
     // Only create stylesheet once with lazy initial state
     // x-ref: https://reactjs.org/docs/hooks-reference.html#lazy-initial-state
     const [styledComponentsStyleSheet] = useState(() => new ServerStyleSheet());