+++ packages/admin/admin/src/utils/test-utils.tsx
@@ -1,7 +1,7 @@
 // Inspired by https://testing-library.com/docs/react-testing-library/setup/#custom-render
 import { createTheme } from "@mui/material";
 import { render, RenderOptions, RenderResult } from "@testing-library/react";
-import React from "react";
+import { ReactElement, ReactNode } from "react";
 import { IntlProvider } from "react-intl";
 
 import { MuiThemeProvider } from "../mui/ThemeProvider";
@@ -10,7 +10,7 @@ import { RouterMemoryRouter } from "../router/MemoryRouter";
 const messages = {};
 const theme = createTheme();
 
-function DefaultWrapper({ children }: { children: React.ReactNode }) {
+function DefaultWrapper({ children }: { children?: ReactNode }) {
     return (
         <IntlProvider locale="en" messages={messages}>
             <MuiThemeProvider theme={theme}>
@@ -20,7 +20,7 @@ function DefaultWrapper({ children }: { children: React.ReactNode }) {
     );
 }
 
-function customRender(ui: React.ReactElement, options?: Omit<RenderOptions, "queries">): RenderResult {
+function customRender(ui: ReactElement, options?: Omit<RenderOptions, "queries">): RenderResult {
     return render(ui, { wrapper: DefaultWrapper, ...options });
 }