+++ packages/admin/blocks-admin/src/blocks/common/AdminComponentRoot.tsx
@@ -1,14 +1,13 @@
 import { Stack, StackBreadcrumbs } from "@comet/admin";
 import { styled } from "@mui/material/styles";
-import * as React from "react";
+import { PropsWithChildren, ReactNode } from "react";
 import { FormattedMessage } from "react-intl";
 
 interface Props {
-    children: React.ReactNode;
-    title?: React.ReactNode;
+    title?: ReactNode;
 }
 
-function AdminComponentRoot(props: Props): React.ReactElement {
+const AdminComponentRoot = (props: PropsWithChildren<Props>) => {
     const { children, title = <FormattedMessage id="comet.blocks" defaultMessage="Blocks" /> } = props;
 
     return (
@@ -27,7 +26,7 @@ function AdminComponentRoot(props: Props): React.ReactElement {
             <ChildrenContainer>{children}</ChildrenContainer>
         </Stack>
     );
-}
+};
 
 export { AdminComponentRoot };