+++ packages/site/cms-site/src/blocks/factories/OneOfBlock.tsx
@@ -1,9 +1,9 @@
-import * as React from "react";
+import { PropsWithChildren } from "react";
 
 import { ErrorHandlerBoundary } from "../../errorHandler/ErrorHandlerBoundary";
 import { SupportedBlocks } from "./types";
 
-interface Props {
+interface Props extends PropsWithChildren {
     data: {
         block?: {
             type: string;
@@ -12,11 +12,10 @@ interface Props {
         };
     };
     supportedBlocks: SupportedBlocks;
-    children?: React.ReactNode;
     className?: string;
 }
 
-export const OneOfBlock: React.FC<Props> = ({ data: { block, ...additionalProps }, supportedBlocks, children, className }) => {
+export const OneOfBlock = ({ data: { block, ...additionalProps }, supportedBlocks, children, className }: Props) => {
     if (!block) {
         return null;
     }