+++ packages/site/cms-site/src/iframebridge/useBlockPreviewFetch.tsx
@@ -1,4 +1,4 @@
-import * as React from "react";
+import { useEffect, useRef } from "react";
 
 import { createFetchInMemoryCache } from "../graphQLFetch/fetchInMemoryCache";
 import { createFetchWithPreviewHeaders, createGraphQLFetch } from "../graphQLFetch/graphQLFetch";
@@ -9,10 +9,10 @@ const cachingFetch = createFetchInMemoryCache(fetch);
 export function useBlockPreviewFetch(graphqlApiUrl: string) {
     const iFrameBridge = useIFrameBridge();
 
-    const graphQLFetchRef = React.useRef(
+    const graphQLFetchRef = useRef(
         createGraphQLFetch(createFetchWithPreviewHeaders(cachingFetch, { includeInvisible: !iFrameBridge.showOnlyVisible }), graphqlApiUrl),
     );
-    React.useEffect(() => {
+    useEffect(() => {
         graphQLFetchRef.current = createGraphQLFetch(
             createFetchWithPreviewHeaders(cachingFetch, { includeInvisible: !iFrameBridge.showOnlyVisible }),
             graphqlApiUrl,