+++ packages/site/cms-site/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @comet/cms-site
 
+## 7.2.1
+
 ## 7.2.0
 
 ### Minor Changes